:recycle: General refactoring and project re-organization
