+++ src/prediction.py
@@ -1,7 +1,7 @@
 import torch
 
 
-def get_prediction(input, model):
-    output = model.forward(input, 1)
+def get_prediction(x, model):
+    output = model.forward(x, 1)
     _, predicted = torch.max(output.data, 1)
     return predicted
